feat(#4): trains permissions

+++ app/controllers/trainController.js
@@ -1,24 +1,31 @@
 const Train = require('../models/Train');
+const User = require('../models/User');
 
-const createTrain = (req, res) => {
-    const train = new Train({
-        name: req.body.name,
-        departureTime: req.body.departureTime,
-        startStation: req.body.startStation,
-        endStation: req.body.endStation
-    });
+const createTrain = async (req, res) => {
+    const currentUser = await User.findOne({_id: req.userId});
 
-    train.save()
-        .then(result => {
-         res.status(201).json(result);
-       })
-        .catch (err => {
-            console.log(err);
-            res.status(500).json({ error: 'Internal Server Error' });
-       });
+    if (currentUser.role == "ADMIN") {
+        const train = new Train({
+            name: req.body.name,
+            departureTime: req.body.departureTime,
+            startStation: req.body.startStation,
+            endStation: req.body.endStation
+        });
+
+        train.save()
+            .then(result => {
+            res.status(201).json(result);
+        })
+            .catch (err => {
+                console.log(err);
+                res.status(500).json({ error: 'Internal Server Error' });
+        });
+    } else {
+        res.status(403).json({ error: 'You do not have the suffisant privilieges to perform this action'})
+    }
 };
 
-const getAllTrains = (req, res)=> {
+const getAllTrains = (req, res) => {
     Train.find()
         .populate('startStation')
         .populate('endStation')
@@ -46,33 +53,45 @@ const getSingleTrain = (req, res) => {
         });
 };
 
-const updateTrain = (req, res) => {
-    const id = req.params.id;
+const updateTrain = async (req, res) => {
+    const currentUser = await User.findOne({_id: req.userId});
 
-    Train.findByIdAndUpdate(id , req.body)
-        .then(result => {
-            if (result) {
-                res.status(200).send('Updated train successfully'); // 200 OK
-            } else {
-                res.status(404).json({ error: 'Train not found' }); // 404 Not Found
-            }
-        })
-        .catch(err => {
-            console.error(err);
-            res.status(500).json({ error: 'Internal Server Error' }); // 500 Internal Server Error
-        });
+    if (currentUser.role == "ADMIN") {
+        const id = req.params.id;
+
+        Train.findByIdAndUpdate(id , req.body)
+            .then(result => {
+                if (result) {
+                    res.status(200).send('Updated train successfully'); // 200 OK
+                } else {
+                    res.status(404).json({ error: 'Train not found' }); // 404 Not Found
+                }
+            })
+            .catch(err => {
+                console.error(err);
+                res.status(500).json({ error: 'Internal Server Error' }); // 500 Internal Server Error
+            });
+    } else {
+        res.status(403).json({ error: 'You do not have the suffisant privilieges to perform this action'})
+    }
 };
 
-const deleteTrain = (req, res) => {
-    const id = req.params.id;
+const deleteTrain = async (req, res) => {
+    const currentUser = await User.findOne({_id: req.userId});
 
-    Train.findByIdAndDelete(id)
-        .then(result => {
-            res.send('Train deleted successful');
-        })
-        .catch(err => {
-            console.log(err);
-        });
+    if (currentUser.role == "ADMIN") {
+        const id = req.params.id;
+
+        Train.findByIdAndDelete(id)
+            .then(result => {
+                res.send('Train deleted successful');
+            })
+            .catch(err => {
+                console.log(err);
+            });
+    } else {
+        res.status(403).json({ error: 'You do not have the suffisant privilieges to perform this action'})
+    }
 };
 
 module.exports = { createTrain , getAllTrains, getSingleTrain, updateTrain, deleteTrain };
